src/com/timbrell/george/exercise2/Library/Person.java
package com.timbrell.george.exercise2.Library;

import java.util.Date;
import java.util.UUID;

public class Person {
    private String name;
    private int age;
    private Date dateOfBirth;
    int uniqueID;

    public Person(String name, int age, Date dateOfBirth) {
        this.name = name;
        this.age = age;
        this.dateOfBirth = dateOfBirth;
    }

    public String getName() {
        return name;
    }

    public int getAge() {
        return age;
    }

    public Date getDateOfBirth() {
        return dateOfBirth;
    }

    public void setName(String name) {
        this.name = name;
    }

    public void setAge(int age) {
        this.age = age;
    }

    public void setDateOfBirth(Date dateOfBirth) {
        this.dateOfBirth = dateOfBirth;
    }
}
